aula 30-08

+++ BibliotecaTurma30/src/main/java/com/bcopstein/endpointsdemo1/Livro.java
@@ -1,4 +1,4 @@
 package com.bcopstein.endpointsdemo1;
 
-public record Livro(int codigo,String titulo,String autor,int ano) {  
+public record Livro(Long codigo,String titulo,String autor,int ano) {  
 }
